Fix/enhancement: Adding HashedFieldBehavior
The behavior hashes configure fields when they exists in POSTed data

Without this check, 'secret fields' will be 'corrupted' for some cases.
For eg: when updating redirect_uri (without specifying existing secret):

$this->Client->save(array(
	'id' => 'foo',
	'redirect_uri' => 'http://mysite.com',
));

will cause 'client_secret' to be populated with a hash of an empty string.

The behavior also implements a beforeFind() method, that removes the
need to manually hash field in $conditions